move block definition to videopress-block
