+++ projects/plugins/jetpack/extensions/blocks/videopress-block/view.js
@@ -11,5 +11,5 @@ import './style.scss';
 
 domReady( function () {
 	// eslint-disable-next-line no-console
-	console.log( __( 'VideoPress init', 'jetpack' ) );
+	console.log( __( 'VideoPress init 2022', 'jetpack' ) );
 } );
