--- projects/plugins/jetpack/extensions/blocks/videopress/view.js
@@ -1,1 +1,0 @@
-import './v6/view';
